Take a list of files as input
It makes more sense to handle it this way than taking in a directory, as
this makes it possible to rebuild only a part of the cards in a
directory.

But it introduces an additional risk of accidentally building a card
twice, once from the .front side and once from the .back side. This
shouldn't pose too much of a problem, though, as anki is quite good at
handling duplicates.