Write python script to replace old shell script
The new script now supports an intermediary build step, so that the
usage in makefiles becomes feasible, and it now directly uses files as
output instead of printing everything to stdout.

Furthermore, the 'qls' command is not supported anymore, as it isn't
really of as much use as previously thought.